requirements.txt
Flask_Bootstrap==3.3.7.1
Werkzeug==1.0.1
colorgram.py==1.2.0
WTForms==2.3.3
Flask==1.1.2
Flask_WTF==0.15.1
